chore(mcp): move gdrive to new architecture
